Resolve "Command Execution"

Closes #6, #1, and #3

See merge request formosa-gpgpu/pocl!5

+++ lib/kernel/formosa/get_global_id.c
@@ -21,11 +21,11 @@
    THE SOFTWARE.
 */
 
-size_t _CL_OVERLOADABLE get_local_size(uint);
-size_t _CL_OVERLOADABLE get_group_id(uint);
-size_t _CL_OVERLOADABLE get_local_id(uint);
+uint64_t _CL_OVERLOADABLE get_local_size(uint);
+uint64_t _CL_OVERLOADABLE get_group_id(uint);
+uint64_t _CL_OVERLOADABLE get_local_id(uint);
 
-size_t _CL_OVERLOADABLE
+uint64_t _CL_OVERLOADABLE
 get_global_id(unsigned int dimindx)
 {
   switch(dimindx)
